requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
opencv-python>=4.8.0
numpy>=1.24.0,<3.0.0
jinja2>=3.1.0
pydantic>=2.7.0
pytest>=8.0.0
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0
kagglehub>=0.3.12
pyyaml>=6.0.0
packaging>=24.0
pymongo>=4.8.0
